+++ lib/Auth/Process/OpaqueSmartID.php
@@ -2,6 +2,7 @@
 
 namespace SimpleSAML\Module\userid\Auth\Process;
 
+use SimpleSAML\Auth\ProcessingChain;
 use SimpleSAML\Auth\ProcessingFilter;
 use SimpleSAML\Auth\State;
 use SimpleSAML\Configuration;
@@ -319,6 +320,7 @@ class OpaqueSmartID extends ProcessingFilter
             $request['UserID'] = [$idValue];
             $request['Attributes'][$this->idAttribute] = [$idValue];
             $request['rciamAttributes']['cuid'] = [$idValue];
+            ProcessingChain::resumeProcessing($request);
             return;
         }
         $this->showError(